Move currency selector to trading screen
Advantage:
- Easier to access
- More visible (especially on mobile)

Disadvantage:
- Always present on screen, users that are only interested in one pair might not want to see it at all times
